Now it sends its buttery voltage

+++ components/battery_controller/include/battery_controller.h
@@ -2,6 +2,6 @@
 #define BATTERY_MONITOR_H
 
 void battery_monitor_init(void);
-float battery_monitor_read_voltage(void);
+int battery_monitor_read_voltage(void);
 
 #endif // BATTERY_MONITOR_H
